refactor: extract copy config logic into useCopyConfig composable
